[svn r97] Patch from Kai Thomsen:
On rename operations, slashes in RISC OS pathnames aren't converted
into dots. So when you have, e. g., two directories `a' and `b' in a
HostFS directory on Linux or Mac OS X, renaming `b' to `a/c' (using
RISC OS) results in `b' moving to the `a' directory and being renamed
to `c', as `/' is the directory separator for those two operating
systems.

--HG--
branch : trunk